requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

# Geospatial
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
geopy==2.4.1

# Firebase/Firestore
firebase-admin==6.3.0
google-cloud-firestore==2.14.0

# Address parsing and normalization
usaddress==0.5.10
# postal - optional, requires libpostal C library

# API clients
overpy==0.7  # OpenStreetMap Overpass API
# arcgis - optional, install separately if needed for ArcGIS FeatureServer

# Data validation
pydantic==2.5.3

# Utilities
tqdm==4.66.1  # Progress bars
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Logging
loguru==0.7.2
